Move fastentrypoints.py to utils/ and add it to MANIFEST.in
